+++ .prettierrc
@@ -1,7 +1,7 @@
 {
   "arrowParens": "always",
   "htmlWhitespaceSensitivity": "ignore",
-  "printWidth": 100,
+  "printWidth": 120,
   "semi": true,
   "singleQuote": true,
   "tabWidth": 2,
